requirements.txt
streamlit==1.42.0
python-dotenv==1.0.1
langchain==0.3.18
libmagic==1.0
langchain-community==0.3.17
tiktoken==0.8.0
langchain-core==0.3.34
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
huggingface-hub==0.28.1
transformers==4.48.3
torch==2.6.0
sentence-transformers==3.4.1
faiss-cpu==1.10.0
openai==1.61.1
docx==0.2.4
python-magic==0.4.27
python-docx==1.1.2
PyPDF2==3.0.1
unstructured==0.16.20
unstructured-client==0.29.0
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1
pickleshare==0.7.5
urllib3==2.3.0
requests==2.32.3
PyYAML==6.0.2
fsspec==2025.2.0
